# [FSharp] Stabilize JIT performance