--- a/fsharp/Program.fs
+++ b/fsharp/Program.fs
@@ -35,49 +35,6 @@ let srcDir = __SOURCE_DIRECTORY__
 [<Literal>]
 let topN = 5
 
-let getRelatedPosts (i: int, tagMap: FrozenDictionary<string, int[]>, taggedPostCount: Span<byte>, posts: Span<Post>) =
-        
-    taggedPostCount.Clear()
-    let post = posts[i]
-    for tag in post.tags do
-        for otherPostIdx in tagMap[tag] do
-            incv &taggedPostCount[otherPostIdx]
-    
-    taggedPostCount[i] <- 0uy // ignore self
-    let top5 = Span.op_Implicit(Array.zeroCreate<TopNItem> topN)
-    let mutable minTags: byte = 0uy
-    let taggedPostLength = uint taggedPostCount.Length
-    
-    //  custom priority queue to find top N
-    let mutable p: int = 0
-    while (uint p < taggedPostLength) do
-        while ((uint p < taggedPostLength) && (taggedPostCount[p] <= minTags)) do
-            p <- p + 1
-        if (uint p < taggedPostLength) then
-            let count: byte = taggedPostCount[p]
-            let mutable upperBound: int = topN - 2
-            while upperBound >= 0 && count > top5[upperBound].count do
-                top5[upperBound + 1] <- top5[upperBound]
-                upperBound <- upperBound - 1
-
-            top5[upperBound + 1] <- { count = count; postId = p }
-            minTags <- top5[topN - 1].count
-        
-        p <- p + 1
-        
-    let topPosts = Array.zeroCreate<Post>(topN)
-    
-    
-    let mutable j = 0
-    while j < 5 do
-        topPosts[j] <- posts[top5[j].postId]
-        j <- j + 1
-        
-    
-    { _id = post._id
-      tags = post.tags
-      related = topPosts }
-        
 let getAllRelated (posts: Span<Post>) =
     let postsCount = posts.Length
     
@@ -98,10 +55,49 @@ let getAllRelated (posts: Span<Post>) =
     
     let allRelatedPosts = Array.zeroCreate<RelatedPosts> postsCount
     let mutable taggedPostCount = stackalloc postsCount
-
+    let mutable top5: Span<TopNItem> = stackalloc topN
     
     for i in 0 .. postsCount - 1 do
-        allRelatedPosts[i] <- getRelatedPosts(i,tagMap, taggedPostCount, posts) 
+        taggedPostCount.Clear()
+        top5.Clear()
+        
+        let post = posts[i]
+        
+        for tag in post.tags do
+            for otherPostIdx in tagMap[tag] do
+                incv &taggedPostCount[otherPostIdx]
+        
+        taggedPostCount[i] <- 0uy // ignore self
+        let mutable minTags: byte = 0uy
+        let taggedPostLength = uint postsCount
+        
+        //  custom priority queue to find top N
+        let mutable p: int = 0
+        while (uint p < taggedPostLength) do
+            while ((uint p < taggedPostLength) && (taggedPostCount[p] <= minTags)) do
+                p <- p + 1
+            if (uint p < taggedPostLength) then
+                let count: byte = taggedPostCount[p]
+               
+                let mutable upperBound: int = topN - 2
+                while upperBound >= 0 && count > top5[upperBound].count do
+                    top5[upperBound + 1] <- top5[upperBound]
+                    upperBound <- upperBound - 1
+
+                top5[upperBound + 1] <- { count = count; postId = p }
+                minTags <- top5[topN - 1].count
+            
+            p <- p + 1
+            
+        let topPosts = Array.zeroCreate<Post>(topN)        
+        
+        let mutable j = 0
+        while j < 5 do
+            topPosts[j] <- posts[top5[j].postId]
+            j <- j + 1
+                
+        allRelatedPosts[i] <- { _id = post._id; tags = post.tags; related = topPosts }
+        
     allRelatedPosts