--- a/fsharp/aot/fsharp.rd.xml
+++ b/fsharp/aot/fsharp.rd.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <Directives>
   <Application>
-    <Assembly Name="FSharp">
+    <Assembly Name="fsharp_aot">
       <Type Name="Program+Post" Dynamic="Required All">
       </Type>
       <Type Name="Program+RelatedPosts" Dynamic="Required All">