Attempt to merge it all back into 1 function, stabilize JIT performance and re-use top5 array via stackalloc Span.

Btw, @jinyus is it possible to make sure the machine that runs the test is update to dotnet 8.0.100, which is now out as of yesterday ?
Thanks in advance !