@jinyus 2023-11-15
> Btw, @jinyus is it possible to make sure the machine that runs the test is update to dotnet 8.0.100, which is now out as of yesterday ?

Just updated the dockerfile to install https://aur.archlinux.org/packages/dotnet-sdk-bin

https://github.com/jinyus/related_post_gen/blob/7d3ed3b44747587021c50dfe9b6dfab8c08980cb/Dockerfile#L28