@jinyus 2023-11-15
This improves JIT but negatively affects AOT. Might have to separate them.

Main : `| F# (AOT) | 27.16 ms| 372.71 ms | 3.23 s`

This PR:
```
F# (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/fsharp
        Processing time (w/o IO): 30.767ms
        Processing time (w/o IO): 30.576ms
        Processing time (w/o IO): 32.8067ms
        Processing time (w/o IO): 30.571ms
        Processing time (w/o IO): 31.7849ms
        Processing time (w/o IO): 31.9095ms
        Processing time (w/o IO): 30.4992ms
        Processing time (w/o IO): 30.6402ms
        Processing time (w/o IO): 30.6363ms
        Processing time (w/o IO): 30.5658ms
        Processing time (w/o IO): 32.1989ms
        Processing time (w/o IO): 30.983ms
        Processing time (w/o IO): 30.687ms
          Time (mean ± σ):     762.3 ms ±   3.2 ms    [User: 736.9 ms, System: 87.3 ms]
          Range (min … max):   756.6 ms … 766.9 ms    10 runs
         
F# (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/fsharp
        Processing time (w/o IO): 435.9567ms
        Processing time (w/o IO): 437.9369ms
        Processing time (w/o IO): 437.0622ms
          Time (mean ± σ):      3.628 s ±  0.006 s    [User: 3.539 s, System: 0.303 s]
          Range (min … max):    3.623 s …  3.632 s    2 runs
         
F# (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/fsharp
        Processing time (w/o IO): 3845.8461ms
        Processing time (w/o IO): 3827.6896ms
        Processing time (w/o IO): 3844.5466ms
          Time (mean ± σ):     16.006 s ±  0.020 s    [User: 17.229 s, System: 0.648 s]
          Range (min … max):   15.992 s … 16.020 s    2 runs
```

Main: `| F# (JIT) | 85.52 ms| 638.39 ms | 5.92 s `
```
F# (JIT):

        Benchmark 1: ./bin/release/net8.0/fsharp
        Processing time (w/o IO): 44.2469ms
        Processing time (w/o IO): 44.2379ms
        Processing time (w/o IO): 44.2257ms
        Processing time (w/o IO): 44.0886ms
        Processing time (w/o IO): 44.026ms
        Processing time (w/o IO): 44.2627ms
        Processing time (w/o IO): 44.2123ms
        Processing time (w/o IO): 44.2238ms
        Processing time (w/o IO): 44.2057ms
        Processing time (w/o IO): 44.0758ms
        Processing time (w/o IO): 44.1271ms
        Processing time (w/o IO): 44.157ms
        Processing time (w/o IO): 43.884ms
          Time (mean ± σ):     927.3 ms ±   7.0 ms    [User: 1027.1 ms, System: 106.9 ms]
          Range (min … max):   916.6 ms … 939.0 ms    10 runs
         
F# (JIT):

        Benchmark 1: ./bin/release/net8.0/fsharp
        Processing time (w/o IO): 647.0812ms
        Processing time (w/o IO): 647.0424ms
        Processing time (w/o IO): 647.474ms
          Time (mean ± σ):      3.429 s ±  0.033 s    [User: 3.450 s, System: 0.250 s]
          Range (min … max):    3.406 s …  3.452 s    2 runs
         
F# (JIT):

        Benchmark 1: ./bin/release/net8.0/fsharp
        Processing time (w/o IO): 5673.3064ms
        Processing time (w/o IO): 5676.1413ms
        Processing time (w/o IO): 5724.2469ms
          Time (mean ± σ):     16.844 s ±  0.016 s    [User: 17.858 s, System: 0.600 s]
          Range (min … max):   16.833 s … 16.855 s    2 runs
```